backend/requirements.txt
# Django Core
Django==5.0.0
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.9

# Redis & Caching
redis==5.0.1
django-redis==5.4.0
hiredis==2.3.2

# Celery & Async
celery==5.3.4
flower==2.0.1

# API Documentation
drf-spectacular==0.27.0

# Environment & Configuration
python-decouple==3.8
python-dotenv==1.0.0

# CORS
django-cors-headers==4.3.1

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Production Server
gunicorn==21.2.0

# Utilities
Pillow==10.1.0